requirements.txt
langgraph
langchain
langchain_cerebras
langchain_community
langsmith

fastapi[standard]
uvicorn[standard]

pydantic

python-dotenv

PyMuPDF

opencv-python
pytesseract
numpy

yt-dlp
pytube

assemblyai

motor

passlib[bycrpt]
python-jose[cryptography]
bcrypt==4.0.1